+++ include/multicam_imu_calib/value_key.hpp
@@ -20,7 +20,7 @@
 
 namespace multicam_imu_calib
 {
-using value_key_t = size_t;
+using value_key_t = int64_t;
 }  // namespace multicam_imu_calib
 
 #endif  // MULTICAM_IMU_CALIB__VALUE_KEY_HPP_
